Revert "Merge branch 'hkenney-ticket/master/2157_remove_facts_dot_d'"
This reverts commit cc414a422de0d773a1012ca57f41f15b4d6caf06, reversing
changes made to 29f8f89c19b2189aa78ab430e94671669cb4716c.

Conflicts:
	README.markdown

Without this patch, there is no facts_dot_d functionality and we don't
have it implemented in Facter 2.0.  This is a problem because Puppet
Enterprise and many users rely on facts.d support.  We're also backwards
compatible with Facter 1.6 in stdlib 3.0 so this is a bug fix.