(MODULES-5003) file_line_does_not_change_multiple_lines_when_one_matches
The exists? method is called to determine the need to call the create and
destroy methods. When the ensure, match, and multiple attributes are defined,
the exists? method needs to return false unless all the lines that match the
match attribute equal the line attribute.

The first commit is minimal, and implements this change.
The second commit normalizes the code, which I needed for comprehension.