(maint) re-send push action to transifex
I rebased release on to master, so changes in master to the readme were not interpretted as changes on release in the payload sent to transifex. This commit should mark the README as modified in Transifex and hopefully they will be able to see the changes, e.g. to_json, to_yaml docs.