(maint) Add deprecation call to `sprintf_hash`
The functions was already documented as being deprecated but an actual
call was missing. This commit adds a call.

Note that the sprintf_hash function is a 4.x function and therefore uses
call_function (all other functions use the deprecated `function_<name>`
convention. They also send Ruby Symbol as the name of the function.