(FM-7388) Fix pick_default_spec assumptions about undef
Before this, the pick_default_spec would fail on Puppet 6.0.0
because on Puppet 6.0.0 arguments given to a function are mapped
to comply with the 3x function calling spec.

The tests asserted that :undef and nil were returned as values (when
they were given as arguments) - however this never happens when the
function is called from the Puppet Language. (This the test was testing
assumptions that were not true).

The spec test is now modified so that when it runs for Puppet 6.0.0 and
later it will do the same mapping on the value as done on the input of
that value to the function before it compares the returned value against
the input. Thus retaining the spirit of the test.