(FM-7388) Fix join_keys_to_values_spec for Puppet 6.0.0
Before this the test would fail on Puppet 6.0.0. This was because
testing logic compensated for hash not having defined order and
it tried to call function in a non API way.

Changing the test to simply use the testing framework fixes this
problem - and since Ruby 1.8 is no longer supported there is no need to
jump through a hoop to sort the result before comparing.