(FM-7388) Fix ensure_resources spec for Puppet 6.0.0
Before this the test would fail because it used `subject_call` and
this ended up in an rspec_puppet compatibility method that
expects two parameters `scope` and `args`. The call was however made
without `scope` and this the arguments where taken as the given
scope - with undefined result.

The use of `call` is supposed to show a deprecation warning (maybe there
was one).

The fix in this commit is to change this to `subject.execute`