(MODULES-8760) Add iterative feature to merge() function
This adds a new feature to the `merge()` function such that it builds a
hash from hashes returned from a lambda when the function is given an
`Iterable` as its only argument.

This adds a 4.x version of merge that is backwards compatible (except it
generates different error messages for argument errors). Since a 4.x
version wins over a 3x the new version will win over the old except for
users that use the old API for calling functions (as expected; this is
fine).