Limit the maximum array size produced by range().
It is trivially easy to accidentally DOS a puppet master by requesting
a range larger than was intended, for example
`range('host00/index.html', 'host25/index.html')`. Because jruby's
implementation of unrolling the range is parallelized, this sort of
thing takes down the whole box.

The solution here is a simple sanity-check on it; we could get fancier
with a new argument specifying a larger size etc. if someone thinks it
needed.