Add correct namespace for start_with function
This function was originally introduced in #1086 but never got a
decision about the correct namespace. This commit adds the stdlib::
namespace to the start_with function after a discussion I had with
Daneel on Slack about this. This will at least align the namespace with
the corresponding end_with function to not confuse people.