(PUP-10653) Remove win32/dir constant usage
For Puppet 7 we are dropping the `win32/dir` dependency as we only used
constants from it, which we replaced with environment variables (see:
https://github.com/puppetlabs/puppet/pull/8314).

This would become breaking when using the stdlib module with Puppet 7,
as the win32 dependencies are no longer provided in the puppet gem.

Replace the usage of the `Dir::COMMON_APPDATA` with the
`ALLUSERSPROFILE` environment variable.