Fix `to_yaml` `options` parameter
Use Ruby 2.5's `transform_keys` method to symbolize the keys in the
`options` parameter before passing to Ruby's `to_yaml` function.

Note, this requires Puppet 6 and above for the required Ruby version to
be available on both the agent (for `puppet apply`) _and_ the server
(Puppetserver 6 ships with JRuby 9.2.x which is Ruby 2.5 compatible).

Also fixed up the docs of the `to_json` function for better consistency
between the two functions.