POC deferrable epp function
This function will detect when it's passed a deferred function and will
effectively defer itself rather than rendering a template directly. It
does this by either returning a deferred inline_epp function with the
template source and variables compiled into the catalot, or by rendering
directly with epp and just returning that output.