(PUP-11752) Fix fqdn_rand_string_spec.rb test
Since Puppet `7.23.0` `fqdn_rand` uses the modern structured
`networking` fact instead of the legacy `fqdn` fact.

In this commit we mock the new fact if using `7.23.0` or later.

This is a rebase of
https://github.com/puppetlabs/puppetlabs-stdlib/pull/1294

and incorporates Aria Li's original change and a variation on Josh
Cooper's suggestion for keeping compatibility with Puppet < `7.23.0`