Allow `deferrable_epp` to return a `Sensitive[String]`
If none of the input variables are `Deferred`, then `deferrable_epp`
returns the result of calling normal `epp`.

The `epp` function however will return a `Sensitive` if any of the
variables it was called with were `Sensitive`.

See https://github.com/puppetlabs/puppet/pull/8330

Adding `Sensitive[String]` to the list of allowed return types fixes
this.