re-add support for loading aliases in yaml files
Ruby 3.1 includes Psych 4.0 which no longer loads aliases by default.
Since aliases are part of the yaml spec, it seems sensible to re-add
support.

We need to read the file first to be compatible with ruby2.7 which does
not support the aliases option on YAML.load_file. Use temp files in
rspec tests to avoid needing to mock read.