Pass calling scope to `stdlib::ensure_packages` from shim
Make the `ensure_packages` shim an Internal function and pass scope to
the namespaced version so as to not change the behaviour of where
packages are contained.

When the function was first ported to the new API, it was discussed that
the existing behaviour might not be 'correct', but changing it would be
a breaking change that might have consequences for many users.

In namespacing the function in 9.0.0 we accidentally created a situation
where the namespaced version worked as before, but the non-namespaced
version, (the shim), now behaved differently.

Fixes #1374