Ignore Puppet's `strict` setting when calling function without namespace
Previously, when a user had the Puppet setting `strict` set to `error` (which
is the default in Puppet 8), a call to one of stdlib's functions via the
deprecated non-namespaced function would cause a hard failure instead of just
logging a warning and calling the real namespaced function.

In this change, all of our shims have been updated to call `deprecation` with
its new third parameter, (`use_strict_setting`), set to `false`. The
non-namespaced versions will now only ever log warnings informing users to
moved to the namespaced versions. It will not raise exceptions even if `strict`
is set to `error`.

This change will make it much easier for users to migrate to stdlib 9 (and to
upgrade modules that now depend on stdlib 9)

Fixes #1373